requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
requests>=2.31.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
pyyaml>=6.0
